requirements.txt
requests
pyyaml
docopt
